water-credit-ai/requirements.txt
requests~=2.32.3
pymongo~=4.7.3
fastapi~=0.111.0
pydantic~=2.7.4
solana~=0.34.2
python-dotenv~=1.0.1
uvicorn~=0.30.1